Merge upstream

+++ scripts/setupTypeScript.js
@@ -22,7 +22,7 @@ const projectRoot = argv[2] || path.join(__dirname, "..")
 // Add deps to pkg.json
 const packageJSON = JSON.parse(fs.readFileSync(path.join(projectRoot, "package.json"), "utf8"))
 packageJSON.devDependencies = Object.assign(packageJSON.devDependencies, {
-  "svelte-check": "^0.1.0",
+  "svelte-check": "^1.0.0",
   "svelte-preprocess": "^4.0.0",
   "@rollup/plugin-typescript": "^4.0.0",
   "typescript": "^3.9.3",
@@ -97,8 +97,18 @@ fs.writeFileSync(tsconfigPath, tsconfig)
 if (!argv[2]) {
   // Remove the script
   fs.unlinkSync(path.join(__filename))
-  // Remove the scripts folder
-  fs.rmdirSync(path.join(__dirname))
+
+  // Check for Mac's DS_store file, and if it's the only one left remove it
+  const remainingFiles = fs.readdirSync(path.join(__dirname))
+  if (remainingFiles.length === 1 && remainingFiles[0] === '.DS_store') {
+    fs.unlinkSync(path.join(__dirname, '.DS_store'))
+  }
+
+  // Check if the scripts folder is empty
+  if (fs.readdirSync(path.join(__dirname)).length === 0) {
+    // Remove the scripts folder
+    fs.rmdirSync(path.join(__dirname))
+  }
 }
 
 // Adds the extension recommendation
